+++ src/Plugin.php
@@ -2,21 +2,25 @@
 
 namespace craft\ckeditor;
 
+use Craft;
 use craft\events\RegisterComponentTypesEvent;
 use craft\services\Fields;
 use yii\base\Event;
 
 /**
  * CKEditor plugin.
- * @method static Plugin getInstance()
  *
+ * @method static Plugin getInstance()
+ * @property-read Settings $settings
+ * @method Settings getSettings()
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
- * @since 1.0
  */
 class Plugin extends \craft\base\Plugin
 {
-    // Public Methods
-    // =========================================================================
+    /**
+     * @inheritdoc
+     */
+    public $hasCpSettings = true;
 
     /**
      * @inheritdoc
@@ -28,5 +32,33 @@ class Plugin extends \craft\base\Plugin
         Event::on(Fields::class, Fields::EVENT_REGISTER_FIELD_TYPES, function(RegisterComponentTypesEvent $e) {
             $e->types[] = Field::class;
         });
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function createSettingsModel()
+    {
+        return new Settings();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function settingsHtml()
+    {
+        return Craft::$app->getView()->renderTemplate('ckeditor/_plugin-settings', [
+            'settings' => $this->getSettings(),
+        ]);
+    }
+
+    /**
+     * Returns the CKEditor build URL.
+     *
+     * @return string
+     */
+    public function getBuildUrl(): string
+    {
+        return Craft::parseEnv($this->getSettings()->buildUrl);
     }
 }