+++ src/assets/field/FieldAsset.php
@@ -19,19 +19,26 @@ class FieldAsset extends AssetBundle
     /**
      * @inheritdoc
      */
-    public function init()
-    {
-        $this->sourcePath = __DIR__.'/dist';
+    public $sourcePath = __DIR__ . '/dist';
 
-        $this->depends = [
-            CpAsset::class,
-            CkeditorAsset::class,
-        ];
+    /**
+     * @inheritdoc
+     */
+    public $depends = [
+        CpAsset::class,
+    ];
 
-        $this->css = [
-            'css/ckeditor-field.min.css',
-        ];
+    /**
+     * @inheritdoc
+     */
+    public $css = [
+        'overrides.css',
+    ];
 
-        parent::init();
-    }
+    /**
+     * @inheritdoc
+     */
+    public $js = [
+        'init.js',
+    ];
 }